html/semantics/embedded-content/the-model-element/content-model-categories.tentative.html

<!DOCTYPE html>
<html>
  <head>
    <link
      rel="help"
      href="https://html.spec.whatwg.org/multipage/embedded-content-other.html#the-model-element"
    />
    <title>Parser tests for the <model> element</title>
  </head>
  <body></body>
  <!-- Include the test harness -->
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script>
    test(() => {
      const modelElem = document.createElement("model");
      assert_true(
        modelElem instanceof HTMLModelElement,
        "<model> element should be an instance of HTMLModelElement"
      );
      assert_equals(
        modelElem.tagName,
        "MODEL",
        "Tag name of <model> element should be 'MODEL'"
      );
      assert_equals(
        modelElem.localName,
        "model",
        "Local name of <model> element should be 'model'"
      );
    }, "Interface test: <model> element is an instance of HTMLModelElement");

    test(() => {
      const container = document.createElement("div");
      container.innerHTML = "<model></model>";
      assert_equals(
        container.childNodes.length,
        1,
        "Flow: <div> should have exactly one child node"
      );
      assert_true(
        container.firstChild instanceof HTMLModelElement,
        "Flow: The child node should be a <model> element"
      );
    }, "Flow content test: <model> parses as flow content");

    test(() => {
      const p = document.createElement("p");
      p.innerHTML = "Start <model></model> End";
      // Expecting three nodes: text, <model>, and text.
      assert_equals(
        p.childNodes.length,
        3,
        "Phrasing: <p> should have three child nodes (text, model, text)"
      );
      assert_equals(
        p.childNodes[1].tagName,
        "MODEL",
        "Phrasing: The second child of <p> should be a <model> element"
      );
    }, "Phrasing content test: <model> parses as phrasing content");

    test(() => {
      const objectElem = document.createElement("object");
      objectElem.innerHTML = "<model></model>";
      assert_equals(
        objectElem.childNodes.length,
        1,
        "Embedded: <object> should have exactly one fallback child"
      );
      assert_equals(
        objectElem.firstChild.tagName,
        "MODEL",
        "Embedded: The fallback child should be a <model> element"
      );
    }, "Embedded content test: <model> parses as embedded content");

    test(() => {
      const container = document.createElement("div");
      container.style.fontSize = "16px"; // Ensure base styling for consistent layout.
      container.innerHTML = "<model>Content</model>";
      document.body.appendChild(container);
      const modelElem = container.querySelector("model");
      const rect = modelElem.getBoundingClientRect();
      assert_true(
        rect.width > 0 || rect.height > 0,
        "<model> element should have a non‑empty bounding box, proving that it renders"
      );
      document.body.removeChild(container);
    }, "Palpable content test: <model> renders as palpable content");
  </script>
</html>
